now the codes are completed cheers, next step to compare the accelerate ratio

+++ blas/level_2.h
@@ -6,20 +6,20 @@ namespace alsm
 {
 	//gemv
 	template <DeviceType D, typename T>
-	void gemv(const stream<D>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, size_t m, size_t n, const T* alpha,
-		const T* A, size_t lda, const T* x, const T* beta, T* y);
+	void gemv(const stream<D>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, int m, int n, const T* alpha,
+		const T* A, int lda, const T* x, const T* beta, T* y);
 #if ALSM_USE_CPU
 	template<>
-	__INLINE__ void gemv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, size_t m, size_t n, const float* alpha,
-		const float* A, size_t lda, const float* x, const float * beta, float* y)
+	__INLINE__ void gemv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, int m, int n, const float* alpha,
+		const float* A, int lda, const float* x, const float * beta, float* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
 		CBLAS_TRANSPOSE trans = static_cast<bool>(in_trans) ? CblasTrans : CblasNoTrans;
 		cblas_sgemv(ord, trans, m, n, *alpha, A, lda, x, 1, *beta, y, 1);
 	}
 	template<>
-	__INLINE__ void gemv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, size_t m, size_t n, const double* alpha,
-		const double* A, size_t lda, const double* x, const double * beta, double* y)
+	__INLINE__ void gemv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, int m, int n, const double* alpha,
+		const double* A, int lda, const double* x, const double * beta, double* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
 		CBLAS_TRANSPOSE trans = static_cast<bool>(in_trans) ? CblasTrans : CblasNoTrans;
@@ -28,15 +28,15 @@ namespace alsm
 #endif
 #if ALSM_USE_GPU
 	template<>
-	__INLINE__ void gemv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, size_t m, size_t n, const float* alpha,
-		const float* A, size_t lda, const float* x, const float * beta, float* y)
+	__INLINE__ void gemv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, int m, int n, const float* alpha,
+		const float* A, int lda, const float* x, const float * beta, float* y)
 	{
 		cublasOperation_t trans = static_cast<bool>(in_trans) ? CUBLAS_OP_T : CUBLAS_OP_N;
 		CUBLAS_CHECK_ERR(cublasSgemv(stream.local_handle, trans, m, n, alpha, A, lda, x, 1, beta, y, 1));
 	}
 	template<>
-	__INLINE__ void gemv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, size_t m, size_t n, const double* alpha,
-		const double* A, size_t lda, const double* x, const double * beta, double* y)
+	__INLINE__ void gemv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixTrans in_trans, MatrixMemOrd in_ord, int m, int n, const double* alpha,
+		const double* A, int lda, const double* x, const double * beta, double* y)
 	{
 
 		cublasOperation_t trans = static_cast<bool>(in_trans) ? CUBLAS_OP_T : CUBLAS_OP_N;
@@ -45,11 +45,11 @@ namespace alsm
 #endif
 	//spmv
 	template <DeviceType D, typename T>
-	__INLINE__ void spmv(const stream<D>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const T* alpha,
+	__INLINE__ void spmv(const stream<D>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const T* alpha,
 		const T* AP, const T* x, const T* beta, T* y);
 #if ALSM_USE_CPU
 	template<>
-	__INLINE__ void spmv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const float* alpha,
+	__INLINE__ void spmv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const float* alpha,
 		const float* Ap, const float* x, const float * beta, float* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
@@ -57,7 +57,7 @@ namespace alsm
 		cblas_sspmv(ord, lower, n, *alpha, Ap, x, 1, *beta, y, 1);
 	}
 	template<>
-	__INLINE__ void spmv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const double* alpha,
+	__INLINE__ void spmv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const double* alpha,
 		const double* Ap, const double* x, const double * beta, double* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
@@ -67,14 +67,14 @@ namespace alsm
 #endif
 #if ALSM_USE_GPU
 	template<>
-	__INLINE__ void spmv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const float* alpha,
+	__INLINE__ void spmv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const float* alpha,
 		const float* Ap, const float* x, const float * beta, float* y)
 	{
 		cublasFillMode_t lower = static_cast<bool>(in_lower) ? CUBLAS_FILL_MODE_LOWER : CUBLAS_FILL_MODE_UPPER;
 		CUBLAS_CHECK_ERR(cublasSspmv(stream.local_handle, lower, n, alpha, Ap, x, 1, beta, y, 1));
 	}
 	template<>
-	__INLINE__ void spmv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const double* alpha,
+	__INLINE__ void spmv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const double* alpha,
 		const double* Ap, const double* x, const double * beta, double* y)
 	{
 		cublasFillMode_t lower = static_cast<bool>(in_lower) ? CUBLAS_FILL_MODE_LOWER : CUBLAS_FILL_MODE_UPPER;
@@ -83,20 +83,20 @@ namespace alsm
 #endif
 	//symv
 	template <DeviceType D, typename T>
-	void symv(const stream<D>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const T* alpha,
-		const T* AP, size_t lda, const T* x, const T* beta, T* y);
+	void symv(const stream<D>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const T* alpha,
+		const T* AP, int lda, const T* x, const T* beta, T* y);
 #if ALSM_USE_CPU
 	template<>
-	__INLINE__ void symv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const float* alpha,
-		const float* Ap, size_t lda, const float* x, const float * beta, float* y)
+	__INLINE__ void symv<DeviceType::CPU, float>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const float* alpha,
+		const float* Ap, int lda, const float* x, const float * beta, float* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
 		CBLAS_UPLO lower = static_cast<bool>(in_lower) ? CblasLower : CblasUpper;
 		cblas_ssymv(ord, lower, n, *alpha, Ap, lda, x, 1, *beta, y, 1);
 	}
 	template<>
-	__INLINE__ void symv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const double* alpha,
-		const double* Ap, size_t lda, const double* x, const double * beta, double* y)
+	__INLINE__ void symv<DeviceType::CPU, double>(const stream<DeviceType::CPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const double* alpha,
+		const double* Ap, int lda, const double* x, const double * beta, double* y)
 	{
 		CBLAS_ORDER ord = static_cast<bool>(in_ord) ? CblasColMajor : CblasRowMajor;
 		CBLAS_UPLO lower = static_cast<bool>(in_lower) ? CblasLower : CblasUpper;
@@ -106,15 +106,15 @@ namespace alsm
 
 #if ALSM_USE_GPU
 	template<>
-	__INLINE__ void symv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const float* alpha,
-		const float* Ap, size_t lda, const float* x, const float * beta, float* y)
+	__INLINE__ void symv<DeviceType::GPU, float>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const float* alpha,
+		const float* Ap, int lda, const float* x, const float * beta, float* y)
 	{
 		cublasFillMode_t lower = static_cast<bool>(in_lower) ? CUBLAS_FILL_MODE_LOWER : CUBLAS_FILL_MODE_UPPER;
 		CUBLAS_CHECK_ERR(cublasSsymv(stream.local_handle, lower, n, alpha, Ap, lda, x, 1, beta, y, 1));
 	}
 	template<>
-	__INLINE__ void symv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, size_t n, const double* alpha,
-		const double* Ap, size_t lda, const double* x, const double * beta, double* y)
+	__INLINE__ void symv<DeviceType::GPU, double>(const stream<DeviceType::GPU>& stream, MatrixMemOrd in_ord, MatrixFillLower in_lower, int n, const double* alpha,
+		const double* Ap, int lda, const double* x, const double * beta, double* y)
 	{
 		cublasFillMode_t lower = static_cast<bool>(in_lower) ? CUBLAS_FILL_MODE_LOWER : CUBLAS_FILL_MODE_UPPER;
 		CUBLAS_CHECK_ERR(cublasDsymv(stream.local_handle, lower, n, alpha, Ap, lda, x, 1, beta, y, 1));
